+++ test/cambium/core_test.clj
@@ -18,11 +18,6 @@
   (testing "Normal scenarios"
     (c/info "hello")
     (c/info {:foo "bar" :baz 10 :qux true} "hello with context")
-    (c/with-nested-context {:extra "context" "data" [1 2 :three 'four]}
-      (is (= (c/get-context) {"extra" "context" "data" "[1 2 :three four]"}))
-      (is (= (c/context-val :extra) "context"))
-      (is (nil? (c/context-val "foo")))
-      (c/info {:foo "bar"} "hello with wrapped context"))
     (c/with-logging-context {:extra "context" "data" [1 2 :three 'four]}
       (is (= (c/get-context) {"extra" "context" "data" "[1 2 :three four]"}))
       (is (= (c/context-val :extra) "context"))
@@ -40,11 +35,6 @@
                     c/destringify-val c/decode-val]
         (c/info "hello")
         (c/info {:foo-k "bar" :baz 10 :qux true} "hello with context")
-        (c/with-nested-context {:extra-k "context" "some-data" [1 2 :three 'four]}
-          (is (= (c/get-context) {"extra_k" "context" "some_data" [1 2 :three 'four]}))
-          (is (= (c/context-val :extra-k) "context"))
-          (is (nil? (c/context-val "foo")))
-          (c/info {:foo "bar"} "hello with wrapped context"))
         (c/with-logging-context {:extra-k "context" "some-data" [1 2 :three 'four]}
           (is (= (c/get-context) {"extra_k" "context" "some_data" [1 2 :three 'four]}))
           (is (= (c/context-val :extra-k) "context"))
@@ -58,10 +48,6 @@
                      :baz :quux}
         context-new {:foo 10
                      :bar :baz}
-        nested-diff {[:foo-fighter :learn-to-fly] {:title "learn to fly"
-                                                   :year 1999}
-                     [:foo-fighter :best-of-you ] {:title "best of you"
-                                                   :year 2005}}
         f (fn
             ([]
               (is (= "bar"  (c/context-val :foo)))
@@ -70,29 +56,12 @@
             ([dummy arg]))]
     (testing "with-raw-mdc"
       (is (nil? (c/context-val :foo)) "Attribute not set must be absent before override")
-      (c/with-nested-context context-old
+      (c/with-logging-context context-old
         (f)
-        (c/with-nested-context context-new
+        (c/with-logging-context context-new
           (is (= "10" (c/context-val :foo)))
           (is (= "quux" (c/context-val :baz)) "Delta context override must not remove non-overridden attributes")
-          (is (= "baz" (c/context-val :bar))))
-        (with-redefs [cambium.core/stringify-val   cambium.core/encode-val
-                      cambium.core/destringify-val cambium.core/decode-val]
-          (c/with-nested-context nested-diff
-            (is (= {"learn-to-fly" {"title" "learn to fly"
-                                    "year" 1999}
-                    "best-of-you"  {"title" "best of you"
-                                    "year" 2005}}
-                  (c/context-val :foo-fighter))
-              "nested map comes out preserved as a map")
-            (is (= {"title" "learn to fly"
-                    "year" 1999}
-                  (c/context-val [:foo-fighter :learn-to-fly]))
-              "deep nested map comes out as a map")
-            (c/with-nested-context {[:foo-fighter :learn-to-fly :year] 2000}
-              (is (= {"title" "learn to fly"
-                      "year" 2000}
-                    (c/context-val [:foo-fighter :learn-to-fly])))))))
+          (is (= "baz" (c/context-val :bar)))))
       (c/with-logging-context context-old
         (f)
         (c/with-logging-context context-new
@@ -102,11 +71,8 @@
       (is (nil? (c/context-val :foo)) "Attribute not set must be absent after restoration"))
     (testing "wrap-raw-mdc"
       (is (nil? (c/context-val :foo)))
-      ((c/wrap-nested-context context-old f))
       ((c/wrap-logging-context context-old f))
-      ((c/wrap-nested-context context-old f) :dummy :arg)
       ((c/wrap-logging-context context-old f) :dummy :arg)
-      ((comp (partial c/wrap-nested-context context-new) (c/wrap-nested-context context-old f)))
       ((comp (partial c/wrap-logging-context context-new) (c/wrap-logging-context context-old f)))
       (is (nil? (c/context-val :foo))))))